Load the defaults value only at "build-time"
The ViperBuilder keeps an internal map of the default key/values pair
set during the definition of the Viper.
At "build-time", after loading the environment, the builder parses the
map and load the defaults. This ensures that the default values are
properly namespaced.
Viper takes care of setting these value at the lowest lever of the
priority "line" ensuring that the key/values defined with a file or the
env are not overwritten.