Ignore error from viper.BindEnv
BindEnv binds a Viper key to a ENV variable; it's a variadic function.

If we look at the signature and the implementation of this function, an
error is returned only if no arguments are passed.

```
func BindEnv(input ...string) error { return v.BindEnv(input...) }
func (v *Viper) BindEnv(input ...string) error {
	var key, envkey string
	if len(input) == 0 {
		return fmt.Errorf("BindEnv missing key to bind to")
	}
```

One option is to be extremely defensive and properly handle the error
anyway, the second option, since it's clear that "cannot happen" (now),
is to trust the API of Viper and simply ignore it.