Providing AppName is explicitly needed to instrument the database
It also avoids a circular dependency in case we would read this value
from the Config object.

The package `instrumentation` can't have a dependency to the package
`configuration` because the `configuration` will have to depend on the
package `instrumentation` when the `instrumentation.Config` will be
"mounted" in the `configuration.Config` struct.

To avoid this dependency the `$APP_ENV` variable is explicitly required
by the function `InstrumentDatabase`.