Add a test for InstrumentAWSSession
This test is checking only the presence of the correct tags and not the
HTTP response code received by the AWS API (HTTP 301 in this case).
Unfortunately we cannot use a mocked S3API in this test.

There is a `s3iface` very useful for mocked-based tests but to test the
actual instrumentation we need to act on and with a "real" AWS session,
that is disjoined from the mocked API.

Ref:
https://docs.aws.amazon.com/sdk-for-go/api/service/s3/s3iface/

As additional reference this is the (non-working and partial) snippet
that could be used to attempt another mocking:

```
import (
	"github.com/aws/aws-sdk-go/aws"
	"github.com/aws/aws-sdk-go/aws/credentials"
	"github.com/aws/aws-sdk-go/aws/session"
	"github.com/aws/aws-sdk-go/service/s3"
	"github.com/aws/aws-sdk-go/service/s3/s3iface"
)

type mockS3Client struct {
	s3iface.S3API
}

func (m *mockS3Client) GetObjectWithContext(aws.Context, *s3.GetObjectInput) (*s3.GetObjectOutput, error) {
	// Just do not return error.
	return nil, nil
}

func Test() {
        s3api := &mockS3Client{
        	sess,            // <---- this cannot be done
        }

        _, err := s3api.GetObjectWithContext(ctx, &s3.GetObjectInput{
        	Bucket: aws.String("test-bucket-name"),
        	Key:    aws.String("//test//file//name"),
        })
}
```