feat: Implement a "native" Sentry hook using the official library
Before this commit, the Logrus hook to Sentry was implemented using
evalphobia/logrus_sentry; this library used the old and unmaintained
getsentry/raven-go. This commit introduces a simple implementation of
the same hook using only the official library from Sentry:
getsentry/sentry-go.

For this implementation these are relevant links:

- How to move from raven-go to sentry-go
  https://docs.sentry.io/platforms/go/migration/
- https://github.com/getsentry/sentry-go
- https://docs.sentry.io/platforms/go/config/
- https://github.com/getsentry/sentry-go/blob/master/MIGRATION.md

The context can be found on Jira: CSVC-478
- https://scribdjira.atlassian.net/browse/CSVC-478