feat: Read APP_VERSION and APP_SERVER_NAME from ENV instead config
Especially a dynamic value as the Release tag, should be read from the
environment and not "hard-coded" in a config file.

The APP_VERSION environment variable is set in the Dockerfile with the
value: "ENV APP_VERSION $GIT_COMMIT_SHORT" and the GIT_COMMIT_SHORT is
configured by the Gitlab pipeline. This ensures an automatic and
consistent versioning of the "release" property in each Sentry notifications.