docs: Update `settings` example to avoid misunderstandings
The prefix `APP_*` is used as namespace for the environment variables
(see viper.go [here][0]) and so it should be avoided as variable prefix
in the settings file.

This can lead to  misunderstandings (see [SERF-325][1]) and generate
false expectations wrt the mapping / relationship between the
environment variable names and configuration files variable names.

[0]: <https://git.lo/microservices/sdk/go-sdk/-/blob/a5d76e2f914758b5553ff205562747d7b337fae9/internal/pkg/configuration/builder/viper.go#L72>
[1]: <https://scribdjira.atlassian.net/browse/SERF-325>